NumericFile Added
Added the NumericFile class, which is a wrapper for numeric I/O. The
class is still in development, but it is already quite functional. It is
quite likely that there are a few bugs present at the moment; these will
be fixed as soon as they are found. This class will most likely have all
of the functionality of the FileWrapper class, plus some additional
convenience functions specific to numeric data processing.